backend/app/mongo.py
import os
from motor.motor_asyncio import AsyncIOMotorClient
from dotenv import load_dotenv

load_dotenv()

MONGODB_URI = os.getenv("MONGODB_URI")
MONGODB_DB = os.getenv("MONGODB_DB", "secure_project")

client = AsyncIOMotorClient(MONGODB_URI)
db = client[MONGODB_DB]
