backend/requirements.txt
# FastAPI & Web Framework
fastapi==0.128.0
uvicorn==0.40.0
pydantic==2.12.5
pydantic-core==2.41.5
email-validator==2.3.0

# Firebase
firebase-admin==7.1.0
google-cloud-firestore==2.23.0
google-cloud-storage==3.8.0
google-auth==2.48.0
google-api-core==2.29.0
google-cloud-core==2.5.0
googleapis-common-protos==1.72.0

# JWT & Security
PyJWT==2.11.0
python-jose==3.5.0
passlib==1.7.4
argon2-cffi==25.1.0
argon2-cffi-bindings==25.1.0
python-multipart==0.0.22
bcrypt==5.0.0
cryptography==46.0.4

# Image Processing & Computer Vision
opencv-python==4.13.0.90
opencv-contrib-python==4.13.0.90
pillow==12.1.0
numpy==2.4.2
scipy==1.17.0

# Facial Recognition
mediapipe==0.10.30
face-recognition==1.3.0
face-recognition-models==0.3.0
dlib==20.0.0

# Object Detection & ML
ultralytics==8.4.9
ultralytics-thop==2.0.18
torch==2.10.0
torchvision==0.25.0

# Utilities
python-dotenv==1.2.1
requests==2.32.5
matplotlib==3.10.8
imutils==0.5.4

# Data Processing & Formatting
click==8.3.1
packaging==26.0
cffi==2.0.0
attrs==25.4.0

# Dependencies
wheel==0.46.3
setuptools==80.10.2
six==1.16.0
python-dateutil==2.9.0.post0
contourpy==1.3.3
fonttools==4.61.1
pycparser==3.0
pyparsing==3.3.2
cycler==0.12.1
kiwisolver==1.4.9
zipp==3.23.0

# Additional dependencies
CacheControl==0.14.4
certifi==2026.1.4
charset-normalizer==3.4.4
colorama==0.4.6
dnspython==2.8.0
ecdsa==0.19.1
filelock==3.20.3
flatbuffers==25.12.19
fsspec==2026.1.0
google-resumable-media==2.8.0
grpcio==1.62.0
grpcio-status==1.62.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jax==0.9.0
jaxlib==0.9.0
Jinja2==3.1.6
MarkupSafe==3.0.3
ml_dtypes==0.5.4
mpmath==1.3.0
msgpack==1.1.2
networkx==3.6.1
opt_einsum==3.4.0
polars==1.37.1
proto-plus==1.27.0
protobuf==4.25.8
psutil==7.2.2
pyasn1==0.6.2
pyasn1-modules==0.4.2
PyYAML==6.0.3
rsa==4.9.1
sounddevice==0.5.5
starlette==0.50.0
sympy==1.14.0
typing-extensions==4.15.0
typing-inspection==0.4.2
urllib3==2.6.3
google-crc32c==1.8.0